refactor trie key adapter
